Added a few casts to get rid of compiler warnings (Visual Studio 64bit).
git-svn-id: svn+ssh://svn.fcdonders.nl/home/svnroot/fieldtrip/trunk@1701 0cf7c7f0-3615-4144-b4e6-68da3bce3cd0

+++ realtime/datasource/siemens/sap2matlab.c
@@ -80,7 +80,7 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[]) {
 	
 	if (nrhs!=1) mexErrMsgTxt("This function needs exactly one (string or uint8) argument.");
 	
-	size = mxGetNumberOfElements(prhs[0]);
+	size = (int) mxGetNumberOfElements(prhs[0]);
 	
 	if ((sizeof(char) == sizeof(mxChar) && mxIsChar(prhs[0])) || mxIsUint8(prhs[0]))  {
 		buffer = (char *) mxGetData(prhs[0]);
